Modified page layouts and further styling
Changed layout to include button inside the container of form so that it
can submit the form, which required further readjustments of related
elements to retain desired styling. Experimented with using span as a
label-like element on the first two input fields to allow a more dynamic
and visually aesthetic styling. However, this removed the label for the
two input fields, which may affect the accessibility of the page.